Reorganize Directory Structure
* Reflects how TileDB-VCF is organized
* Rename `SOMAArrayReader` to `TileDBSOMAArray`
