+++ libtiledbsoma/src/cli/cli.cc
@@ -30,7 +30,8 @@
  * This file is currently a sandbox for C++ API experiments
  */
 
-#include <tiledbsoma/tiledbsoma>
+#include "arrow_adapter.h"
+#include "soma/tiledb_soma_array.h"
 
 using namespace tiledbsoma;
 
@@ -44,12 +45,12 @@ void test_sdf(const std::string& uri) {
     // config["sm.mem.total_budget"] = "1118388608";
 
     // Read all values from the obs array
-    auto obs = SOMAArrayReader::open(uri + "/obs", "obs");
+    auto obs = TileDBSOMAArray::open(uri + "/obs", "obs");
     obs->submit();
     auto obs_data = obs->read_next();
 
     // Read all values from the var array
-    auto var = SOMAArrayReader::open(uri + "/ms/RNA/var", "var");
+    auto var = TileDBSOMAArray::open(uri + "/ms/RNA/var", "var");
     var->submit();
     auto var_data = var->read_next();
 
@@ -59,7 +60,7 @@ void test_sdf(const std::string& uri) {
     }
 
     // Read all values from the X/data array
-    auto x_data = SOMAArrayReader::open(
+    auto x_data = TileDBSOMAArray::open(
         uri + "/ms/RNA/X/data", "X/data", config);
     x_data->submit();
 
@@ -79,7 +80,7 @@ void test_sdf(const std::string& uri) {
 namespace tdbs = tiledbsoma;
 void test_arrow(const std::string& uri) {
     const std::vector<std::string>& colnames{"n_counts", "n_genes", "louvain"};
-    auto obs = tdbs::SOMAArrayReader::open(uri, "", {}, colnames);
+    auto obs = tdbs::TileDBSOMAArray::open(uri, "", {}, colnames);
     obs->submit();
     // Getting next batch:  std::optional<std::shared_ptr<ArrayBuffers>>
     auto obs_data = obs->read_next();
