+++ libtiledbsoma/src/utils/common.h
