+++ libtiledbsoma/src/utils/logger_public.h
