+++ libtiledbsoma/src/utils/spdlog_with_R.h
