+++ libtiledbsoma/src/utils/stats.h
