+++ libtiledbsoma/src/utils/util.h
@@ -1,5 +1,5 @@
 /**
- * @file   util.cc
+ * @file   util.h
  *
  * @section LICENSE
  *
@@ -27,50 +27,46 @@
  *
  * @section DESCRIPTION
  *
- * This file defines utilities.
+ * This file defines the utility functions
  */
 
-#include "tiledbsoma/util.h"
+#ifndef UTIL_H
+#define UTIL_H
+
+#include <regex>
+#include <span/span.hpp>
+#include <stdexcept>  // for windows: error C2039: 'runtime_error': is not a member of 'std'
 
 namespace tiledbsoma::util {
 
+using VarlenBufferPair =
+    std::pair<std::vector<std::byte>, std::vector<uint64_t>>;
+
 template <typename T>
-VarlenBufferPair to_varlen_buffers(std::vector<T> data, bool arrow) {
-    size_t nbytes = 0;
-    for (auto& elem : data) {
-        nbytes += elem.size();
-    }
+VarlenBufferPair to_varlen_buffers(std::vector<T> data, bool arrow = true);
 
-    std::vector<std::byte> result(nbytes);
-    std::vector<uint64_t> offsets(data.size() + 1);
-    size_t offset = 0;
-    size_t idx = 0;
-
-    for (auto& elem : data) {
-        std::memcpy(result.data() + offset, elem.data(), elem.size());
-        offsets[idx++] = offset;
-
-        offset += elem.size();
-    }
-    offsets[idx] = offset;
-
-    // Remove extra arrow offset when creating buffers for TileDB write
-    if (!arrow) {
-        offsets.pop_back();
-    }
-
-    return {result, offsets};
+template <class T>
+std::vector<T> to_vector(const tcb::span<T>& s) {
+    return std::vector<T>(s.begin(), s.end());
 }
 
-template VarlenBufferPair to_varlen_buffers(
-    std::vector<std::string>, bool arrow);
+/**
+ * @brief Check if the provided URI is a TileDB Cloud URI.
+ *
+ * @param uri URI to check
+ * @return true URI is a TileBD Cloud URI
+ * @return false URI is not a TileBD Cloud URI
+ */
+bool is_tiledb_uri(std::string_view uri);
 
-bool is_tiledb_uri(std::string_view uri) {
-    return uri.find("tiledb://") == 0;
-}
+/**
+ * @brief Remove all trailing '/' from URI.
+ *
+ * @param uri URI
+ * @return std::string URI without trailing '/'
+ */
+std::string rstrip_uri(std::string_view uri);
 
-std::string rstrip_uri(std::string_view uri) {
-    return std::regex_replace(std::string(uri), std::regex("/+$"), "");
-}
+}  // namespace tiledbsoma::util
 
-};  // namespace tiledbsoma::util
+#endif
